backend/requirements-prod.txt
# ============================================================================
# PRODUCTION RUNTIME REQUIREMENTS (Lean Docker Image)
# ============================================================================
# This file contains ONLY packages needed to run the API server in Docker.
# For training, evaluation, and data processing, use requirements.txt instead.
#
# Why separate?
#   - requirements.txt includes ~500MB of ML training packages (shap, prophet,
#     geopandas, matplotlib, statsmodels) that are NOT needed at API runtime.
#   - This file produces a Docker image ~1-2GB smaller.
#
# Usage:
#   Docker production:  pip install -r requirements-prod.txt
#   Local development:  pip install -r requirements.txt
#   Training pipeline:  pip install -r requirements.txt
# ============================================================================

# ============================================================================
# WEB FRAMEWORK & API
# ============================================================================
Flask==3.0.3
flask-cors==6.0.0
Werkzeug==3.1.5

# Production Servers
gunicorn==23.0.0

# ============================================================================
# HTTP & EXTERNAL APIs
# ============================================================================
requests==2.32.4
urllib3==2.6.3
aiohttp==3.13.3
aiohttp[speedups]==3.13.3
tenacity==9.1.2

# ============================================================================
# DATA PROCESSING (Runtime - model inference only)
# ============================================================================
# PINNED for ML model reproducibility - must match training environment
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.8.0
joblib==1.4.2
scipy==1.14.1

# ============================================================================
# SCHEDULING & BACKGROUND TASKS
# ============================================================================
APScheduler==3.10.4
celery==5.4.0

# ============================================================================
# CONFIGURATION & ENVIRONMENT
# ============================================================================
python-dotenv==1.0.1

# ============================================================================
# DATABASE & ORM
# ============================================================================
SQLAlchemy==2.0.35
alembic==1.13.3
pg8000==1.31.5
psycopg2-binary==2.9.10

# ============================================================================
# SECURITY & AUTHENTICATION
# ============================================================================
cryptography==45.0.3
itsdangerous==2.2.0
bcrypt==4.3.0
PyJWT[crypto]==2.10.1  # >=2.10.1 required by supabase-auth

# ============================================================================
# INPUT VALIDATION & SANITIZATION
# ============================================================================
validators==0.34.0
bleach==6.2.0
email-validator==2.2.0

# ============================================================================
# RATE LIMITING & CACHING
# ============================================================================
Flask-Limiter==4.1.1
Flask-Session==0.8.0
redis==5.2.0
limits==3.13.0
packaging==24.2

# ============================================================================
# PERFORMANCE OPTIMIZATION
# ============================================================================
flask-compress==1.15
Flask-Caching==2.3.0

# ============================================================================
# API DOCUMENTATION
# ============================================================================
flasgger==0.9.7.1
PyYAML==6.0.2

# ============================================================================
# MONITORING, LOGGING & ERROR TRACKING
# ============================================================================
python-json-logger==3.2.1
prometheus-flask-exporter==0.23.1
sentry-sdk[flask]==2.18.0

# ============================================================================
# PROGRESS TRACKING
# ============================================================================
tqdm==4.67.1

# ============================================================================
# WEATHER DATA SOURCES (used by runtime services)
# ============================================================================
meteostat==1.6.8

# Open-Meteo
openmeteo-requests==1.2.0
requests-cache==1.2.1
retry-requests==2.0.0

# ============================================================================
# EXCEL & PDF (used by upload/export API routes)
# ============================================================================
openpyxl==3.1.5
reportlab==4.2.5

# ============================================================================
# SERVER-SENT EVENTS (SSE)
# ============================================================================
sse-starlette==2.2.1
flask-sse==1.0.0

# ============================================================================
# GRAPHQL
# ============================================================================
graphene==3.4.3
# graphql-server removed - unused (graphql-core provides graphql_sync directly)
graphql-core==3.2.5

# ============================================================================
# GOOGLE CLOUD / EARTH ENGINE (lazy-loaded in google_weather_service)
# ============================================================================
earthengine-api==1.4.3
google-cloud-bigquery==3.27.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-cloud-storage==2.19.0
protobuf==5.29.5
